# fl_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v0/spiele": {
      "get": {
        "summary": "List Spiele",
        "description": "List Spiele matching the given filters.\n\nOmitting `saison_id` returns the CURRENT season. `saison_phase` accepts `playoffs` as an alias\nfor \"any phase except gruppenphase\", and `ausstehend` includes today.",
        "operationId": "get_spiele_api_v0_spiele_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "saison_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saison Id"
            }
          },
          {
            "name": "saison_phase",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "playoffs",
                  "type": "string"
                },
                {
                  "enum": [
                    "gruppenphase",
                    "achtelfinale",
                    "viertelfinale",
                    "halbfinale",
                    "finale"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saison Phase"
            }
          },
          {
            "name": "spiel_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ausstehend",
                    "vergangen",
                    "heute",
                    "abgesagt",
                    "unbekannt"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Spiel Status"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1024,
              "minimum": 1,
              "default": 1024,
              "title": "Limit"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "datum",
                "uhrzeit",
                "spiel_nr",
                "saison_phase"
              ],
              "type": "string",
              "default": "datum",
              "title": "Sort By"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpieleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/spiele/{spiel_id}": {
      "get": {
        "summary": "One Spiel",
        "description": "Return one match by its id.\n\nNo season is resolved and no status derived: the match carries its own `saison_id`, and\n`spiel_status` is a property of a query rather than of a match.",
        "operationId": "get_spiel_api_v0_spiele__spiel_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spiel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spiel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLSpieleSingleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a Spiel",
        "description": "Update one Spiel and resolve the season's bracket.\n\nThe payload is written wholesale: an omitted field is overwritten. A result can fill or empty the\nslots below it, each named in `advanced_to`. `dry_run=true` answers the same and writes nothing.",
        "operationId": "patch_spiel_data_api_v0_spiele__spiel_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spiel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spiel Id"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Report what this payload would move and destroy, and write nothing",
              "default": false,
              "title": "Dry Run"
            },
            "description": "Report what this payload would move and destroy, and write nothing"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPatchSpielDataPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLPatchSpielDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FLBracketFaultGruppe": {
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "gruppe_too_small",
              "tie_unresolved"
            ],
            "title": "Reason"
          },
          "spiel_id": {
            "type": "string",
            "title": "Spiel Id"
          },
          "spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spiel Nr"
          },
          "gruppe": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ],
            "title": "Gruppe"
          },
          "platz": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Platz"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "spiel_id",
          "spiel_nr",
          "gruppe",
          "platz"
        ],
        "title": "FLBracketFaultGruppe",
        "description": "One bracket slot whose `gruppe` reference names a placing no standing will hand it.\n\n`gruppe_too_small` is a typo and the slot is left alone; `tie_unresolved` is a played-out group\nthe chain cannot separate, so the slot IS emptied."
      },
...
      "FLBracketFaultQuelle": {
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "spiel_missing",
              "reference_cycle"
            ],
            "title": "Reason"
          },
          "spiel_id": {
            "type": "string",
            "title": "Spiel Id"
          },
          "spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spiel Nr"
          },
          "quelle_spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quelle Spiel Nr"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "spiel_id",
          "spiel_nr",
          "quelle_spiel_nr"
        ],
        "title": "FLBracketFaultQuelle",
        "description": "One bracket slot whose `spiel` reference names a match that cannot state an outcome.\n\nBoth leave the slot as it stands, and neither is reachable through the write path. A cycle is\nreported on every fixture it reaches."
      },
      "FLBracketFaultSpiel": {
        "properties": {
          "reason": {
            "type": "string",
            "const": "same_team",
            "title": "Reason"
          },
          "spiel_id": {
            "type": "string",
            "title": "Spiel Id"
          },
          "spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spiel Nr"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "spiel_id",
          "spiel_nr"
        ],
        "title": "FLBracketFaultSpiel",
        "description": "One fixture whose two references resolve to the SAME club.\n\nIt survives the write-path rule, which keys a source by identity, so two DIFFERENT sources\nnaming one club pass. The fixture keeps its stored sides."
      },
...
      "FLPatchSpielDataPayload": {
        "properties": {
          "is_canceled": {
            "type": "boolean",
            "title": "Is Canceled"
          },
          "team1": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielTeamField"
              },
              {
                "type": "null"
              }
            ]
          },
          "team2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielTeamField"
              },
              {
                "type": "null"
              }
            ]
          },
          "team1_quelle": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleGruppe"
                  },
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "gruppe": "#/components/schemas/FLSpielQuelleGruppe",
                    "spiel": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Team1 Quelle"
          },
          "team2_quelle": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleGruppe"
                  },
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "gruppe": "#/components/schemas/FLSpielQuelleGruppe",
                    "spiel": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Team2 Quelle"
          },
          "elfmeterschiessen": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielElfmeterschiessen"
              },
              {
                "type": "null"
              }
            ]
          },
          "datum": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datum"
          },
          "uhrzeit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uhrzeit"
          },
          "ort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielOrtField"
              },
              {
                "type": "null"
              }
            ]
          },
          "schiedsrichter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielSchiedsrichterField"
              },
              {
                "type": "null"
              }
            ]
          },
          "notiz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notiz"
          }
        },
        "type": "object",
        "required": [
          "is_canceled",
          "team1",
          "team2",
          "team1_quelle",
          "team2_quelle",
          "elfmeterschiessen",
          "datum",
          "uhrzeit",
          "ort",
          "schiedsrichter",
          "notiz"
        ],
        "title": "FLPatchSpielDataPayload"
      },
      "FLPatchSpielDataResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "advanced_to": {
            "items": {
              "$ref": "#/components/schemas/FLSpielAdvancement"
            },
            "type": "array",
            "title": "Advanced To"
          },
          "released_sides": {
            "items": {
              "$ref": "#/components/schemas/FLSpielReleasedSide"
            },
            "type": "array",
            "title": "Released Sides"
          },
          "bracket_faults": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/FLBracketFaultGruppe"
                },
                {
                  "$ref": "#/components/schemas/FLBracketFaultQuelle"
                },
                {
                  "$ref": "#/components/schemas/FLBracketFaultSpiel"
                },
                {
                  "$ref": "#/components/schemas/FLBracketFaultOccupant"
                }
              ],
              "discriminator": {
                "propertyName": "reason",
                "mapping": {
                  "disqualified_occupant": "#/components/schemas/FLBracketFaultOccupant",
                  "gruppe_too_small": "#/components/schemas/FLBracketFaultGruppe",
                  "reference_cycle": "#/components/schemas/FLBracketFaultQuelle",
                  "same_team": "#/components/schemas/FLBracketFaultSpiel",
                  "spiel_missing": "#/components/schemas/FLBracketFaultQuelle",
                  "tie_unresolved": "#/components/schemas/FLBracketFaultGruppe"
                }
              }
            },
            "type": "array",
            "title": "Bracket Faults"
          }
        },
        "type": "object",
        "title": "FLPatchSpielDataResponse",
        "description": "What `patch_spiel_data` returns: every fixture it moved, and what that cost.\n\nBoth lists report writes the caller did not ask for, each carrying the result it destroyed."
      },
...
      "FLSpielAdvancement": {
        "properties": {
          "spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spiel Nr"
          },
          "voided_ergebnis": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]+:[0-9]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided Ergebnis"
          },
          "voided_elfmeterschiessen": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielElfmeterschiessen"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "spiel_nr",
          "voided_ergebnis",
          "voided_elfmeterschiessen"
        ],
        "title": "FLSpielAdvancement",
        "description": "One fixture the bracket resolution rewrote, and the result that rewrite destroyed.\n\nBoth voided fields are `None` where a slot merely filled from empty, so \"was anything destroyed\"\nis a null check."
      },
      "FLSpielElfmeterschiessen": {
        "properties": {
          "team1": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Team1"
          },
          "team2": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Team2"
          }
        },
        "type": "object",
        "required": [
          "team1",
          "team2"
        ],
        "title": "FLSpielElfmeterschiessen",
        "description": "The penalty shoot-out that settled a knockout fixture whose goals finished level.\n\nNo `sieger`: a second statement of the same fact could contradict the counts, and no validator\ncould express that it must not (`docs/backend/spec.md :: I25`)."
      },
      "FLSpielJoined": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "team1": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielTeamFieldJoined"
              },
              {
                "type": "null"
              }
            ]
          },
          "team2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielTeamFieldJoined"
              },
              {
                "type": "null"
              }
            ]
          },
          "team1_quelle": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleGruppe"
                  },
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "gruppe": "#/components/schemas/FLSpielQuelleGruppe",
                    "spiel": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Team1 Quelle"
          },
          "team2_quelle": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleGruppe"
                  },
                  {
                    "$ref": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "gruppe": "#/components/schemas/FLSpielQuelleGruppe",
                    "spiel": "#/components/schemas/FLSpielQuelleSpiel"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Team2 Quelle"
          },
          "datum": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datum"
          },
          "uhrzeit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uhrzeit"
          },
          "ort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielOrtField"
              },
              {
                "type": "null"
              }
            ]
          },
          "schiedsrichter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielSchiedsrichterField"
              },
              {
                "type": "null"
              }
            ]
          },
          "ergebnis": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]+:[0-9]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ergebnis"
          },
          "elfmeterschiessen": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielElfmeterschiessen"
              },
              {
                "type": "null"
              }
            ]
          },
          "spieltag_id": {
            "type": "string",
            "title": "Spieltag Id"
          },
          "spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spiel Nr"
          },
          "is_canceled": {
            "type": "boolean",
            "title": "Is Canceled"
          },
          "saison_phase": {
            "type": "string",
            "enum": [
              "gruppenphase",
              "achtelfinale",
              "viertelfinale",
              "halbfinale",
              "finale"
            ],
            "title": "Saison Phase"
          },
          "saison_id": {
            "type": "string",
            "maxLength": 4,
            "minLength": 4,
            "title": "Saison Id"
          },
          "notiz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notiz"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team1",
          "team2",
          "team1_quelle",
          "team2_quelle",
          "datum",
          "uhrzeit",
          "ort",
          "schiedsrichter",
          "ergebnis",
          "elfmeterschiessen",
          "spieltag_id",
          "spiel_nr",
          "is_canceled",
          "saison_phase",
          "saison_id"
        ],
        "title": "FLSpielJoined",
        "description": "One fixture as an ENDPOINT serves it: the stored document plus each side's joined season state.\n\n**A subclass, and the direction must not be flipped**: the stored shape is the one a write can\nreach, so it stays free of derived fields."
      },
      "FLSpielOrtField": {
        "properties": {
          "spielort_id": {
            "type": "string",
            "title": "Spielort Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "maps_link": {
            "type": "string",
            "minLength": 1,
            "title": "Maps Link"
          },
          "mietpreis": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Mietpreis"
          }
        },
        "type": "object",
        "required": [
          "spielort_id",
          "name",
          "maps_link",
          "mietpreis"
        ],
        "title": "FLSpielOrtField"
      },
      "FLSpielQuelleGruppe": {
        "properties": {
          "type": {
            "type": "string",
            "const": "gruppe",
            "title": "Type"
          },
          "gruppe": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ],
            "title": "Gruppe"
          },
          "platz": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Platz"
          }
        },
        "type": "object",
        "required": [
          "type",
          "gruppe",
          "platz"
        ],
        "title": "FLSpielQuelleGruppe",
        "description": "A slot fed by the group phase: the team finishing `platz` in `gruppe`."
      },
      "FLSpielQuelleSpiel": {
        "properties": {
          "type": {
            "type": "string",
            "const": "spiel",
            "title": "Type"
          },
          "spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spiel Nr"
          },
          "ausgang": {
            "type": "string",
            "enum": [
              "sieger",
              "verlierer"
            ],
            "title": "Ausgang"
          }
        },
        "type": "object",
        "required": [
          "type",
          "spiel_nr",
          "ausgang"
        ],
        "title": "FLSpielQuelleSpiel",
        "description": "A slot fed by an earlier fixture: the side that came out of match `spiel_nr` as `ausgang`."
      },
      "FLSpielReleasedSide": {
        "properties": {
          "spiel_nr": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spiel Nr"
          },
          "side": {
            "type": "string",
            "enum": [
              "team1",
              "team2"
            ],
            "title": "Side"
          },
          "team_name": {
            "type": "string",
            "minLength": 1,
            "title": "Team Name"
          },
          "voided_ergebnis": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]+:[0-9]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided Ergebnis"
          },
          "voided_elfmeterschiessen": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLSpielElfmeterschiessen"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "spiel_nr",
          "side",
          "team_name",
          "voided_ergebnis",
          "voided_elfmeterschiessen"
        ],
        "title": "FLSpielReleasedSide",
        "description": "One side another fixture gave up so a team can play this Spieltag.\n\n`team_name` rather than an id: the message quoting this has no `spiele` list to join against,\nand the rename fan-out keeps the copy it reads fresh."
      },
      "FLSpielSchiedsrichterField": {
        "properties": {
          "schiedsrichter_id": {
            "type": "string",
            "title": "Schiedsrichter Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "payment": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Payment"
          }
        },
        "type": "object",
        "required": [
          "schiedsrichter_id",
          "name",
          "payment"
        ],
        "title": "FLSpielSchiedsrichterField"
      },
      "FLSpielTeamField": {
        "properties": {
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "tore": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Tore"
          },
          "shorthand": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Shorthand"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "name",
          "tore",
          "shorthand"
        ],
        "title": "FLSpielTeamField",
        "description": "One side of a fixture as the DOCUMENT stores it and the admin PATCH writes it back.\n\nNothing joined belongs here: the payload is written back wholesale, so a field added reaches the\ndocument on the next edit."
      },
...
      "FLSpieleListResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "spiele": {
            "items": {
              "$ref": "#/components/schemas/FLSpielJoined"
            },
            "type": "array",
            "title": "Spiele"
          }
        },
        "type": "object",
        "required": [
          "spiele"
        ],
        "title": "FLSpieleListResponse"
      },
      "FLSpieleSingleResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "spiel": {
            "$ref": "#/components/schemas/FLSpielJoined"
          }
        },
        "type": "object",
        "required": [
          "spiel"
        ],
        "title": "FLSpieleSingleResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}